Fix Sphinx warnings for partition_counts by renaming the module and dropping the type alias.
