+++ src/pixelator/pna/analysis/segmentation/__init__.py
@@ -3,6 +3,6 @@
 Copyright © 2026 Pixelgen Technologies AB.
 """
 
-from pixelator.pna.analysis.segmentation.partition_counts import partition_counts
+from pixelator.pna.analysis.segmentation.partition import partition_counts
 
 __all__ = ["partition_counts"]
